working on l2norm

+++ src/blas-level1.c
@@ -267,7 +267,7 @@ gauss_Error gauss_vec_scale(
                 dst->data.flt,
                 a->data.flt,
                 size,
-                *(double *)scalar
+                *(float *)scalar
             );
             break;
